Added -Wextra to compilation options, fixed warnings
Increased the rigor of compile-time checks, and fixed many
small problems brought up by this.

The most significant change was on {read,write}_{float,sfloat}
functions at src/util/bytelib.h, where exponents had unsigned
type but could be negative. It was passing unit tests using GCC
but another compiler might do differently.

+++ src/communication/fsm.c
@@ -479,7 +479,7 @@ char *fsm_get_current_state_name(FSM *fsm)
  */
 char *fsm_state_to_string(fsm_states state)
 {
-	if (state >= fsm_state_disconnected && state < fsm_state_size) {
+	if ((int) state >= fsm_state_disconnected && state < fsm_state_size) {
 		return fsm_state_strings[state];
 	}
 
@@ -494,7 +494,7 @@ char *fsm_state_to_string(fsm_states state)
 char *fsm_event_to_string(fsm_events evt)
 {
 
-	if (evt >= fsm_evt_ind_transport_connection && evt < fsm_evt_size) {
+	if ((int) evt >= fsm_evt_ind_transport_connection && evt < fsm_evt_size) {
 		return fsm_event_strings[evt];
 	}
